# docs/vajra_api_openapi.yaml
openapi: 3.1.0
info:
  title: Vajra V2 Platform API
  version: 1.0.0-draft
  description: Azure-ready API contract for Vajra predictive maintenance, RAG, agents, HITL and operations.
servers:
  - url: /api/v1
paths:
...
  /telemetry/window:
    get:
      summary: Return an asset telemetry window
      parameters:
        - in: query
          name: asset_id
          required: true
          schema: {type: string}
        - in: query
          name: row_index
          schema: {type: integer, minimum: 0}
        - in: query
          name: window
          schema: {type: integer, minimum: 1, maximum: 2000, default: 250}
      responses:
        '200':
          description: Telemetry window
          content:
            application/json:
              schema:
                type: object
                properties:
                  metadata: {$ref: '#/components/schemas/ResponseMetadata'}
                  records:
                    type: array
                    items: {type: object, additionalProperties: true}
...
  /predictions:
    post:
      summary: Run Vajra predictive inference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        '200':
          description: Prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
...
components:
  schemas:
    ResponseMetadata:
      type: object
      required: [request_id, generated_at, app_version]
      properties:
        request_id: {type: string, format: uuid}
        generated_at: {type: string, format: date-time}
        app_version: {type: string}
        predictive_model_version: {type: [string, 'null']}
        llm_model_version: {type: [string, 'null']}
        prompt_version: {type: [string, 'null']}
        rag_index_version: {type: [string, 'null']}
...
    PredictionRequest:
      oneOf:
        - type: object
          required: [source, row_index]
          properties:
            source: {const: replay}
            row_index: {type: integer, minimum: 0}
            scenario_run_id: {type: string}
        - type: object
          required: [source, asset_id, timestamp, features]
          properties:
            source: {const: features}
            asset_id: {type: string}
            timestamp: {type: string, format: date-time}
            features: {type: object, additionalProperties: {type: number}}
    PredictionResponse:
      type: object
      required: [metadata, prediction]
      properties:
        metadata: {$ref: '#/components/schemas/ResponseMetadata'}
        prediction:
          type: object
          required: [alarm_id, risk_level, risk_score, predicted_fault]
          properties:
            alarm_id: {type: string}
            timestamp_index: {type: integer}
            row_index: {type: integer}
            asset_name: {type: string}
            is_alert: {type: boolean}
            anomaly_probability: {type: number}
            anomaly_threshold: {type: number}
            predicted_fault: {type: string}
            predicted_stand: {type: string}
            fault_confidence: {type: number}
            secondary_faults: {type: array, items: {type: object, additionalProperties: true}}
            predicted_rul_band: {type: string}
            proxy_rul_observations: {type: number}
            proxy_rul_shifts: {type: number}
            trend_risk: {type: number}
            risk_score: {type: number}
            risk_level: {type: string}
            health_index: {type: number}
            evidence: {type: array, items: {type: string}}
            similar_historical_cases: {type: array, items: {type: object, additionalProperties: true}}
            disclaimer: {type: string}
...